# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for GoPOS Backend",
        "title": "GoPOS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/reports/gross-profit": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve summary and product breakdown of gross profit within a specific date range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get Gross Profit Report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date in YYYY-MM-DD format (default: today)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date in YYYY-MM-DD format (default: today)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved gross profit report",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/transactions/{id}/refund": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Refund specific items from a completed transaction, restore stock, deduct shift cash, and log the action.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Process a partial or full transaction refund",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Refund Data Payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RefundInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Refund processed successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.RefundInput": {
            "type": "object",
            "required": [
                "items",
                "reason"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/handlers.RefundItemInput"
                    }
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "handlers.RefundItemInput": {
            "type": "object",
            "required": [
                "product_id",
                "qty_refunded"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "qty_refunded": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}